+++ packages/plugin-table/web/src/tiptap/consts.ts
@@ -4,3 +4,5 @@ export const TRANSACTION_META_KEYS = {
   SELECT_COLUMN: 'select-column',
   SELECT_ROW: 'select-row',
 };
+
+export const TABLE_PLUGIN_MENU_BUTTON_ID = 'table_plugin_menu';